backend/requirements.txt
fastapi
uvicorn[standard]
pyswisseph
httpx
pydantic[email]
